test: Move comparison tests for AttributeDto to AttributeDtoTest
